requirements/base.txt
Django==3.1.5
django-debug-toolbar==2.2
django-filter==2.2.0
django-stubs==1.5.0
django-timezone-field==4.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
Markdown==3.2.1
mypy==0.770
mypy-extensions==0.4.3
Pillow==7.1.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-memcached==1.59
django-ckeditor==5.8.0
PyYAML==5.3.1
django-admin-sortable2
django-constance
django-picklefield
django-rest-swagger
drf-yasg
